[feat] 폴더링 수정
