+++ src/main/java/com/wanted_pre/wanted_pre_onboarding_backend/sevice/recruitment/RecruitmentService.java
@@ -1,15 +1,16 @@
-package com.wanted_pre.wanted_pre_onboarding_backend.sevice;
+package com.wanted_pre.wanted_pre_onboarding_backend.sevice.recruitment;
 
 import com.wanted_pre.wanted_pre_onboarding_backend.common.exception.enums.ErrorCode;
 import com.wanted_pre.wanted_pre_onboarding_backend.common.exception.model.NotFoundException;
-import com.wanted_pre.wanted_pre_onboarding_backend.controller.dto.request.RecruitmentCreateRequest;
-import com.wanted_pre.wanted_pre_onboarding_backend.controller.dto.request.RecruitmentUpdateRequest;
-import com.wanted_pre.wanted_pre_onboarding_backend.controller.dto.response.DetailRecruitmentReadResponse;
-import com.wanted_pre.wanted_pre_onboarding_backend.controller.dto.response.RecruitmentReadResponse;
+import com.wanted_pre.wanted_pre_onboarding_backend.controller.recruitment.dto.request.RecruitmentCreateRequest;
+import com.wanted_pre.wanted_pre_onboarding_backend.controller.recruitment.dto.request.RecruitmentUpdateRequest;
+import com.wanted_pre.wanted_pre_onboarding_backend.controller.recruitment.dto.response.DetailRecruitmentReadResponse;
+import com.wanted_pre.wanted_pre_onboarding_backend.controller.recruitment.dto.response.RecruitmentReadResponse;
 import com.wanted_pre.wanted_pre_onboarding_backend.domain.enterprise.Enterprise;
 import com.wanted_pre.wanted_pre_onboarding_backend.domain.recruitment.Recruitment;
 import com.wanted_pre.wanted_pre_onboarding_backend.repository.EnterpriseRepository;
 import com.wanted_pre.wanted_pre_onboarding_backend.repository.RecruitmentRepository;
+import com.wanted_pre.wanted_pre_onboarding_backend.sevice.enterprise.EnterpriseService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -24,13 +25,13 @@ import java.util.stream.Collectors;
 @RequiredArgsConstructor
 public class RecruitmentService {
 
-    private final EnterpriseRepository enterpriseRepository;
     private final RecruitmentRepository recruitmentRepository;
+    private final EnterpriseService enterpriseService;
 
     @Transactional
     public void postRecruitment(final RecruitmentCreateRequest request) {
         //1. 회사 존재 여부 찾기
-        Enterprise enterprise = enterpriseRepository.findById(request.enterpriseId()).orElseThrow(() -> new NotFoundException(ErrorCode.ENTERPRISE_NOT_FOUND_EXCEPTION));
+        Enterprise enterprise = enterpriseService.getEnterprise(request.enterpriseId());
 
         //2. 등록하기
         Recruitment recruitment = new Recruitment(request.reward(), request.position(), request.detail(), request.skill(), enterprise);
@@ -40,8 +41,7 @@ public class RecruitmentService {
     @Transactional
     public void putRecruitment(final Long recruitmentId, final RecruitmentUpdateRequest request) {
         //1. 채용 공고 존재 여부 찾기
-        Recruitment recruitment = recruitmentRepository.findById(recruitmentId).orElseThrow(() -> new NotFoundException(ErrorCode.RECRUITMENT_NOT_FOUND_EXCEPTION));
-
+        Recruitment recruitment = getRecruitment(recruitmentId);
         //2. 내용 업데이트
         recruitment.update(request.reward(), request.position(), request.detail(), request.skill());
         recruitmentRepository.save(recruitment);
@@ -50,8 +50,7 @@ public class RecruitmentService {
     @Transactional
     public void deleteRecruitment(final Long recruitmentId){
         //1. 채용 공고 존재 여부 찾기
-        Recruitment recruitment = recruitmentRepository.findById(recruitmentId).orElseThrow(() -> new NotFoundException(ErrorCode.RECRUITMENT_NOT_FOUND_EXCEPTION));
-
+        Recruitment recruitment = getRecruitment(recruitmentId);
         //2. 삭제
         recruitmentRepository.delete(recruitment);
     }
@@ -85,8 +84,7 @@ public class RecruitmentService {
     }
 
     public DetailRecruitmentReadResponse getDetailRecruitment(final Long recruitmentId){
-        Recruitment recruitment = recruitmentRepository.findById(recruitmentId).orElseThrow(() -> new NotFoundException(ErrorCode.RECRUITMENT_NOT_FOUND_EXCEPTION));
-
+        Recruitment recruitment = getRecruitment(recruitmentId);
         List<Long> anotherRecruitmentIdList = recruitmentRepository.findRecruitmentIdsByEnterpriseId(recruitment.getEnterprise().getId(), recruitmentId);
 
         DetailRecruitmentReadResponse detailRecruitmentReadResponse = new DetailRecruitmentReadResponse(
@@ -104,6 +102,9 @@ public class RecruitmentService {
         return detailRecruitmentReadResponse;
     }
 
+    public Recruitment getRecruitment(final Long recruitmentId) {
+        return recruitmentRepository.findById(recruitmentId).orElseThrow(() -> new NotFoundException(ErrorCode.RECRUITMENT_NOT_FOUND_EXCEPTION));
+    }
 
     private List<RecruitmentReadResponse> getRecruitmentResponse(List<Recruitment> recruitmentList) {
         List<RecruitmentReadResponse> recruitmentReadResponseList = recruitmentList.stream().map( recruitment -> {
